Fix forge url line numbers bug
Currently, the commit ref for a yanked selection is taken to be the most
recent commit hash for the selected region (rather than for the file as a
whole).

This leads to incorrectly generated forge URLs, since changes elsewhere in a
file can impact line numbers for the yanked range.

This patch restores the previous behavior of using the most recent commit ref
for the file as a whole.

Possibilities for future development:

- Stash before yanking
- Check to see if the current commit has been pushed
- Warn before yanking uncommitted or unpushed changes